+++ dotnet/Eyes.Selenium.DotNet/Fluent/IPathNodeValue.cs
@@ -1,0 +1,6 @@
+namespace Applitools.Selenium
+{
+    public interface IPathNodeValue
+    {
+    }
+}