+++ dotnet/Eyes.Selenium.DotNet/Fluent/RegionLocator.cs
@@ -1,0 +1,10 @@
+namespace Applitools.Selenium
+{
+    public class RegionLocator : TargetPathLocator
+    {
+        public RegionLocator(TargetPathLocator parent, IPathNodeValue value)
+            : base(parent, value)
+        {
+        }
+    }
+}