+++ dotnet/Eyes.Selenium4.DotNet/Fluent/ElementReference.cs
@@ -1,0 +1,14 @@
+using OpenQA.Selenium;
+
+namespace Applitools.Selenium
+{
+    public class ElementReference : IPathNodeValue
+    {
+        public IWebElement Element { get; }
+
+        public ElementReference(IWebElement element)
+        {
+            Element = element;
+        }
+    }
+}