+++ dotnet/Eyes.Selenium4.DotNet/Fluent/TargetPathLocator.cs
@@ -1,0 +1,18 @@
+namespace Applitools.Selenium
+{
+    public class TargetPathLocator
+    {
+        protected internal TargetPathLocator Parent { get; }
+        protected internal  IPathNodeValue Value { get; }
+
+        public TargetPathLocator()
+        {
+        }
+
+        public TargetPathLocator(TargetPathLocator parent, IPathNodeValue value)
+        {
+            Parent = parent;
+            Value = value;
+        }
+    }
+}