+++ dotnet/download_cores.ps1
@@ -1,10 +1,6 @@
 #!/usr/bin/env pwsh
-Param
-(
-    [String] [Parameter(Mandatory=$true)] $CORE_VERSION = "3.1.1"
-)
 
-$version = $CORE_VERSION
+$version = ${env:CORE_VERSION}
 wget -nv "https://github.com/applitools/eyes.sdk.javascript1/releases/download/js%2Fcore%40$version/core-alpine"
 wget -nv "https://github.com/applitools/eyes.sdk.javascript1/releases/download/js%2Fcore%40$version/core-linux"
 wget -nv "https://github.com/applitools/eyes.sdk.javascript1/releases/download/js%2Fcore%40$version/core-linux-arm64"