+++ dotnet/CHANGELOG.md
@@ -1,3 +1,85 @@
+## [Eyes.Images 3.30] - 2023-07-09
+### Added
+- New emulated devices:
+  * Galaxy S21
+  * Galaxy S21 Ultra
+### Removed
+- Emulated devices:
+  * Galaxy S21 5G
+  * Galaxy S22 Ultra 5G
+
+## [Eyes.Selenium 3.59] - 2023-07-09
+### Updated
+- Match to latest Eyes.Images
+
+## [Eyes.Appium 5.55] - 2023-07-09
+### Updated
+- Match to latest Eyes.Selenium
+
+## [Eyes.Selenium4 3.59] - 2023-07-09
+### Updated
+- Match to latest Eyes.Images
+
+## [Eyes.Playwright 1.7] - 2023-07-09
+### Updated
+- Match to latest Eyes.Images
+
+## [Eyes.Selenium 3.58] - 2023-07-09
+### Added
+- New emulated devices:
+  * Galaxy S21
+  * Galaxy S21 Ultra
+### Removed
+- Emulated devices:
+  * Galaxy S21 5G
+  * Galaxy S22 Ultra 5G
+
+## [Eyes.Appium 5.54] - 2023-07-09
+### Updated
+- Match to latest Eyes.Selenium
+
+## [Eyes.Selenium4 3.58] - 2023-07-09
+### Added
+- New emulated devices:
+  * Galaxy S21
+  * Galaxy S21 Ultra
+### Removed
+- Emulated devices:
+  * Galaxy S21 5G
+  * Galaxy S22 Ultra 5G
+
+## [Eyes.Image.Core 3.5.1] - 2023-07-06
+### Updated
+- Server Core to 3.5.1
+
+## [Eyes.Images 3.29] - 2023-07-06
+### Updated
+- Match to latest Eyes.Image.Core
+
+## [Eyes.Selenium 3.57] - 2023-07-06
+### Added
+- Shadow DOM support
+- New emulated devices:
+  * Galaxy S21 5G
+  * Galaxy S22 Ultra
+  * Galaxy S21 Ultra 5G
+
+## [Eyes.Appium 5.53] - 2023-07-06
+### Updated
+- Match to latest Eyes.Selenium
+
+## [Eyes.Selenium4 3.57] - 2023-07-06
+### Added
+- Shadow DOM support
+- New emulated devices:
+  * Galaxy S21 5G
+  * Galaxy S22 Ultra
+  * Galaxy S21 Ultra 5G
+
+## [Eyes.Playwright 1.6] - 2023-07-06
+### Updated
+- Match to latest Eyes.Images
+
 ## [Eyes.Image.Core 3.3.1] - 2023-07-02
 ### Updated
 - Server Core to 3.3.1